requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
gitpython>=3.1.0
pydantic>=2.0.0
pytest>=8.0.0

# Video generation dependencies
replicate>=0.20.0
pillow>=10.0.0
ffmpeg-python>=0.2.0
httpx>=0.25.0

# Optional: OpenAI for alternative LLM
openai>=1.3.0
